[#154747783] Cleanup
